@@ -1,6 +1,6 @@
 ////// valNumber validator
 
-import { default as _isNaN } from 'lodash/isNaN';
+import { default as _isNaN } from 'lodash-es/isNaN.js';
 
 import { type, check as isNumber } from '../validation/isNumber';
 
