@@ -2,7 +2,7 @@
 // Utility for testing validation and validator modules
 //
 
-import _ from 'lodash';
+import _ from 'lodash-es';
 import { types } from '../../src/lib/types';
 import { qualifiers, DEFAULT_QUALIFIER } from '../../src/lib/qualifiers';
 import * as util from '../../src/lib/util';
